refactor: Mark image file input as unrequired
At the beggining the client was always required to pass image in the
request body. Now it is not required anymore. If the client does not
pass an image, a placeholder image will be used instead.

In case of updating an item, if the client does not pass an image, the
old image will be used instead, if it exists, otherwise a placeholder.